refactor: reorganize diff code
